[APPC-3538] moved Log to module
