+++ app/src/main/java/com/asfoundation/wallet/analytics/BackendEventLogger.java
@@ -4,7 +4,7 @@ import cm.aptoide.analytics.AnalyticsManager;
 import cm.aptoide.analytics.AnalyticsManager.Action;
 import cm.aptoide.analytics.EventLogger;
 import com.asf.wallet.BuildConfig;
-import com.asfoundation.wallet.util.Log;
+import com.appcoins.wallet.core.utils.common.Log;
 import io.reactivex.schedulers.Schedulers;
 import java.util.Map;
