+++ app/src/main/java/com/asfoundation/wallet/analytics/RakamAnalytics.kt
@@ -12,7 +12,7 @@ import com.asfoundation.wallet.logging.RakamReceiver
 import com.asfoundation.wallet.promo_code.repository.PromoCode
 import com.asfoundation.wallet.promo_code.repository.PromoCodeLocalDataSource
 import com.asfoundation.wallet.promotions.model.PromotionsModel
-import com.asfoundation.wallet.util.Log
+import com.appcoins.wallet.core.utils.common.Log
 import com.google.android.gms.common.ConnectionResult
 import com.google.android.gms.common.GoogleApiAvailability
 import dagger.hilt.android.qualifiers.ApplicationContext
