+++ app/src/main/java/com/asfoundation/wallet/analytics/RakamEventLogger.kt
@@ -2,7 +2,7 @@ package com.asfoundation.wallet.analytics
 
 import cm.aptoide.analytics.AnalyticsManager
 import cm.aptoide.analytics.EventLogger
-import com.asfoundation.wallet.util.Log
+import com.appcoins.wallet.core.utils.common.Log
 import io.rakam.api.Rakam
 import org.json.JSONException
 import org.json.JSONObject
