+++ app/src/main/java/com/asfoundation/wallet/logging/DebugReceiver.kt
@@ -2,7 +2,7 @@ package com.asfoundation.wallet.logging
 
 import com.appcoins.wallet.commons.LogReceiver
 import com.asf.wallet.BuildConfig
-import com.asfoundation.wallet.util.Log
+import com.appcoins.wallet.core.utils.common.Log
 
 class DebugReceiver : LogReceiver {
