+++ app/src/main/java/com/asfoundation/wallet/ui/SendActivity.kt
@@ -9,15 +9,13 @@ import android.view.View
 import android.widget.EditText
 import android.widget.ImageButton
 import android.widget.Toast
-import androidx.activity.viewModels
 import androidx.lifecycle.ViewModelProvider
-import androidx.lifecycle.ViewModelProviders
 import com.asf.wallet.R
 import com.asfoundation.wallet.C
 import com.asfoundation.wallet.router.Result
 import com.asfoundation.wallet.ui.barcode.BarcodeCaptureActivity
 import com.asfoundation.wallet.ui.iab.IabActivity
-import com.asfoundation.wallet.util.Log.Companion.e
+import com.appcoins.wallet.core.utils.common.Log.Companion.e
 import com.asfoundation.wallet.viewmodel.SendViewModel
 import com.asfoundation.wallet.viewmodel.SendViewModelFactory
 import com.google.android.gms.common.api.CommonStatusCodes
