+++ app/src/main/java/com/asfoundation/wallet/ui/camera/CameraSourcePreview.java
@@ -24,7 +24,7 @@ import android.view.SurfaceView;
 import android.view.ViewGroup;
 import androidx.annotation.RequiresPermission;
 import com.asfoundation.wallet.ui.barcode.CameraResultListener;
-import com.asfoundation.wallet.util.Log;
+import com.appcoins.wallet.core.utils.common.Log;
 import com.google.android.gms.common.images.Size;
 import java.io.IOException;
