+++ app/src/main/java/com/asfoundation/wallet/ui/iab/MergedAppcoinsPresenter.kt
@@ -9,7 +9,7 @@ import com.asfoundation.wallet.entity.TransactionBuilder
 import com.asfoundation.wallet.ui.iab.MergedAppcoinsFragment.Companion.APPC
 import com.asfoundation.wallet.ui.iab.MergedAppcoinsFragment.Companion.CREDITS
 import com.appcoins.wallet.core.utils.common.CurrencyFormatUtils
-import com.asfoundation.wallet.util.Log
+import com.appcoins.wallet.core.utils.common.Log
 import com.appcoins.wallet.core.utils.common.WalletCurrency
 import com.asfoundation.wallet.util.isNoNetworkException
 import com.asfoundation.wallet.wallets.usecases.GetWalletInfoUseCase
