+++ app/src/main/java/com/asfoundation/wallet/util/KS.java
@@ -8,6 +8,7 @@ import android.content.Intent;
 import android.security.keystore.KeyGenParameterSpec;
 import android.security.keystore.KeyProperties;
 import android.security.keystore.UserNotAuthenticatedException;
+import com.appcoins.wallet.core.utils.common.Log;
 import com.asf.wallet.R;
 import com.asfoundation.wallet.entity.ServiceErrorException;
 import java.io.ByteArrayOutputStream;
